+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* Added a macro to roll weather and a setting to disable automatically rolling weather when a new day begins
+
 ## [0.0.1] - 2023-01-05
 
 ### Added